+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [8.1.2]
 ### Changed
 - EMC for [Mystical Agriculture] `cognizant_dust` to `16384` because it is a rarer mob drop
 
@@ -770,7 +772,8 @@ Added support for 18 more mods
 - Added support for Extended Crafting
     - Combintation Core, Compressor, Ender Crafter and Tiered Tables
 
-[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.1.1...HEAD
+[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.1.2...HEAD
+[8.1.2]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.1.1...v8.1.2
 [8.1.1]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.1.0...v8.1.1
 [8.1.0]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.0.3...v8.1.0
 [8.0.3]: https://github.com/TagnumElite/ProjectE-Integration/compare/v8.0.2...v8.0.3